reorganize the tests a little
